# cmake: allow building tests in unity mode