--- a/tests/libtest/lib677.c
+++ b/tests/libtest/lib677.c
@@ -28,7 +28,7 @@
 #include "memdebug.h"
 
 static const char cmd[] = "A1 IDLE\r\n";
-static char buf[1024];
+static char testbuf[1024];
 
 CURLcode test(char *URL)
 {
@@ -101,9 +101,9 @@ CURLcode test(char *URL)
           pos = 0;
         }
       }
-      else if(pos < (ssize_t)sizeof(buf)) {
+      else if(pos < (ssize_t)sizeof(testbuf)) {
         CURLcode ec;
-        ec = curl_easy_recv(curl, buf + pos, sizeof(buf) - pos, &len);
+        ec = curl_easy_recv(curl, testbuf + pos, sizeof(testbuf) - pos, &len);
         if(ec == CURLE_AGAIN) {
           continue;
         }
@@ -122,7 +122,7 @@ CURLcode test(char *URL)
   }
 
   if(state) {
-    fwrite(buf, pos, 1, stdout);
+    fwrite(testbuf, pos, 1, stdout);
     putchar('\n');
   }