--- a/tests/server/resolve.c
+++ b/tests/server/resolve.c
@@ -21,6 +21,13 @@
  * SPDX-License-Identifier: curl
  *
  ***************************************************************************/
+/* Hack for Unity mode */
+#ifdef HEADER_CURL_MEMDEBUG_H
+#undef HEADER_CURL_MEMDEBUG_H
+#undef freeaddrinfo
+#undef getaddrinfo
+#endif
+
 #include "server_setup.h"
 
 /* Purpose