--- a/tests/server/resolve.c
+++ b/tests/server/resolve.c
@@ -21,6 +21,10 @@
  * SPDX-License-Identifier: curl
  *
  ***************************************************************************/
+#ifndef CURL_NO_GETADDRINFO_OVERRIDE
+#define CURL_NO_GETADDRINFO_OVERRIDE
+#endif
+
 #include "server_setup.h"
 
 /* Purpose
@@ -125,11 +129,9 @@ int main(int argc, char *argv[])
     hints.ai_family = use_ipv6 ? PF_INET6 : PF_INET;
     hints.ai_socktype = SOCK_STREAM;
     hints.ai_flags = 0;
-    /* Use parenthesis around functions to stop them from being replaced by
-       the macro in memdebug.h */
-    rc = (getaddrinfo)(host, "80", &hints, &ai);
+    rc = getaddrinfo(host, "80", &hints, &ai);
     if(rc == 0)
-      (freeaddrinfo)(ai);
+      freeaddrinfo(ai);
   }
 #else
   if(use_ipv6) {